Introduces maintenance concern
