+++ src/api/.rubocop_todo.yml
@@ -537,6 +537,7 @@ Naming/PredicateName:
     - 'app/models/package.rb'
     - 'app/models/patchinfo.rb'
     - 'app/models/project.rb'
+    - 'app/models/concerns/project_maintenance.rb'
     - 'app/models/repository.rb'
     - 'app/models/user.rb'
     - 'app/models/user_basic_strategy.rb'
@@ -940,6 +941,7 @@ Rails/InverseOf:
     - 'app/models/message.rb'
     - 'app/models/package.rb'
     - 'app/models/project.rb'
+    - 'app/models/concerns/project_maintenance.rb'
     - 'app/models/rating.rb'
     - 'app/models/repository.rb'
     - 'app/models/review.rb'
